requirements.txt
fastapi==0.115.5
pydantic==2.9.2
uvicorn==0.32.0
pymongo==4.9.2
python-dotenv==1.0.1
PyJWT==2.9.0
passlib==1.7.4